refactor: backend full refactor (separated controller and service)

+++ paint-backend/src/main/java/com/example/paint_backend/exception/ShapeNotFoundException.java
@@ -1,14 +1,14 @@
 package com.example.paint_backend.exception;
 
 public class ShapeNotFoundException extends RuntimeException {
-    private final int shapeId;
+    private final Long shapeId;
 
-    public ShapeNotFoundException(int shapeId) {
+    public ShapeNotFoundException(Long shapeId) {
         super("Shape not found with ID: " + shapeId);
         this.shapeId = shapeId;
     }
 
-    public int getShapeId() {
+    public Long getShapeId() {
         return shapeId;
     }
 }
